# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  responses.EditionResponse:
    properties:
      acquired_at:
        description: YYYY-MM-DD
        type: string
      book_id:
        type: string
      copy_count:
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration_seconds:
        type: integer
      edition_name:
        type: string
      format:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      isbn_10:
        type: string
      isbn_13:
        type: string
      language:
        type: string
      narrator:
        type: string
      page_count:
        type: integer
      publish_date:
        description: YYYY-MM-DD
        type: string
      publisher:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Librarium
    url: https://github.com/fireball1725/librarium
  description: Self-hosted library tracking API. All protected endpoints require a
    Bearer JWT in the Authorization header.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  termsOfService: http://localhost/
  title: Librarium API
  version: 26.0.0-dev
paths:
...
  /libraries/{library_id}/books/{book_id}/editions:
    get:
      description: Returns all editions (physical/digital/audio) for a book.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.EditionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List editions of a book
      tags:
      - editions
    post:
      consumes:
      - application/json
      description: Adds a new edition to an existing book.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition details
        in: body
        name: body
        required: true
        schema:
          properties:
            acquired_at:
              type: string
            copy_count:
              type: integer
            description:
              type: string
            duration_seconds:
              type: integer
            edition_name:
              type: string
            format:
              type: string
            is_primary:
              type: boolean
            isbn_10:
              type: string
            isbn_13:
              type: string
            language:
              type: string
            narrator:
              type: string
            page_count:
              type: integer
            publish_date:
              type: string
            publisher:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.EditionResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an edition
      tags:
      - editions
  /libraries/{library_id}/books/{book_id}/editions/{edition_id}:
    delete:
      description: Permanently deletes an edition from a book.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an edition
      tags:
      - editions
    put:
      consumes:
      - application/json
      description: Replaces an edition's metadata.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      - description: Edition details
        in: body
        name: body
        required: true
        schema:
          properties:
            acquired_at:
              type: string
            copy_count:
              type: integer
            description:
              type: string
            duration_seconds:
              type: integer
            edition_name:
              type: string
            format:
              type: string
            is_primary:
              type: boolean
            isbn_10:
              type: string
            isbn_13:
              type: string
            language:
              type: string
            narrator:
              type: string
            page_count:
              type: integer
            publish_date:
              type: string
            publisher:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.EditionResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an edition
      tags:
      - editions
  /libraries/{library_id}/books/{book_id}/editions/{edition_id}/files:
    post:
      consumes:
      - multipart/form-data
      description: Attaches an ebook or audiobook file to an edition.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      - description: Book file (epub, pdf, mp3, m4b, etc.)
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload edition file
      tags:
      - editions
  /libraries/{library_id}/books/{book_id}/editions/{edition_id}/files/{file_id}:
    delete:
      description: Removes a specific file from an edition.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      - description: File UUID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove edition file
      tags:
      - editions
    get:
      description: Streams a specific file attached to a digital edition.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      - description: File UUID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Download edition file
      tags:
      - editions
  /libraries/{library_id}/books/{book_id}/editions/{edition_id}/files/link:
    post:
      consumes:
      - application/json
      description: Links a file that already exists on the server (within a storage
        location) to an edition.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link existing file to edition
      tags:
      - editions
  /libraries/{library_id}/books/{book_id}/editions/{edition_id}/files/link-upload:
    post:
      consumes:
      - application/json
      description: Links a file that already exists in the server's default upload
        directory to an edition.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Book UUID
        in: path
        name: book_id
        required: true
        type: string
      - description: Edition UUID
        in: path
        name: edition_id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Link file from upload path to edition
      tags:
      - editions
...
  /libraries/{library_id}/browse-uploads:
    get:
      description: Lists files and subdirectories within the server's configured ebook
        or audiobook upload path.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: ebook or audiobook
        in: query
        name: format
        required: true
        type: string
      - description: 'Relative sub-path (default: root)'
        in: query
        name: path
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Browse default upload directory
      tags:
      - editions
...
  /libraries/{library_id}/storage-locations/{location_id}/browse:
    get:
      description: Lists files and subdirectories within a storage location at the
        given path.
      parameters:
      - description: Library UUID
        in: path
        name: library_id
        required: true
        type: string
      - description: Storage location UUID
        in: path
        name: location_id
        required: true
        type: string
      - description: 'Relative sub-path (default: root)'
        in: query
        name: path
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Browse storage location directory
      tags:
      - storage-locations
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"